requirements.txt
https://github.com/boto/boto/tarball/develop
https://github.com/marcelnicolay/asyncdynamo/tarball/0.2.6
tornado
simplejson
python-memcached
